Don't regenerate IDs immediately after inserting subtitle(s)
When M-i is held down for a few seconds, the ID regeneration can sometimes end
up in an infinite(?) loop for some reason. Only scheduling the regeneration to
be done in 100ms and cancelling that call if another subtitle is inserted within
the next 100ms seems to fix that issue.

This also makes inserting subtitles much quicker.