Move subtitles in active region as a unit
Usually, it's illegal to make the start time larger than the stop time or to
make the stop time larger than the next subtitle's start time.  But if we move
subtitles in a region, we expect their time stamps not to change relatively to
each other.  That means we must allow negative durations and negative subtitle
spacing when moving the active region.

But we only allow it in the active region.  We don't want the leading
subtitle (i.e. the last subtitle in the active region when moving forward and
the first subtitle in the active region when moving backward) to be able to
overlap with it's neighbour.  So we move the leading subtitle first, applying
all the usual limitations, and check how far it actually moved.  Then we can
move the rest of the subtitles by the same amount.