Adopt REUSE standard for copyright and licensing
Add copyright notices and SPDX license headers to all files. This is for
conformance to the REUSE specification (https://reuse.software/spec/),
which standardises licensing information in order to help both humans
and computers in free software licensing audits.

Add a 'contributions' section to the README to explain this convention
for new contributors to the project.

Move all the file local variables in the 'tests/' directory to the
bottom of the files in order to make space for the REUSE headers. See
GNU Emacs manual page 'Local Variables in Files' for an explanation of
this syntax.