subed-split-subtitle: Use offset or text fraction
* subed/subed-common.el (subed-split-subtitle): Add an optional
argument OFFSET, which can be the number of milliseconds to split at
after the beginning (if positive) or before the end (if negative) of
the current subtitle. If it is not specified and MPV is not playing or
if the playback position is outside the current subtitle, use the
timestamp proportional to the point's position between start and stop
timestamp of the current subtitle.

If called interactively, calling it with one prefix argument
prompts for the offset in milliseconds.  Calling it with two
prefix arguments uses the relative position of the point even if MPV
is playing.