New command subed-narrow-to-times
* subed/subed-common.el (subed-narrow-to-times): New function makes it
easier to focus on part of the subtitle file. This looks for the first
subtitle on or after the start and the last subtitle on or before the
end. It should also handle the case when there are no subtitles for those
precise moments.
(subed--string-to-msecs): Utility function for specifying times in
timestamps, milliseconds, or seconds.