Adopt REUSE standard for copyright and licensing
Add copyright notices and SPDX license headers for all files. This is
for conformance to the [REUSE](https://reuse.software/spec/)
specification, which standardises licensing information in order to help
both humans and computers in free software licensing audits.

Add a AUTHORS.org file for contributors to optionally credit themselves
in.

Add a 'contributions' section to the README to explain the REUSE headers
and AUTHORS.org file to new contributors to the project.