Create derived modes for the different types of subtitles.
This makes it easier to manually set a buffer's mode in case a file
uses a different extension.

* README.org: Add note about derived modes.
* subed/subed-ass.el (subed-ass-mode): New derived mode.
(auto-mode-alist): Handle .ass files.
* subed/subed-srt.el (subed-srt-mode): New derived mode.
(auto-mode-alist): Handle .srt files.
* subed/subed-vtt.el (subed-vtt-mode): New derived mode.
(auto-mode-alist): Handle .vtt files.
* subed/subed.el (subed--init): New optional format parameter
specifies the format to load.
(subed-auto-find-video-maybe): Moved from subed--init to a function so
that it can be in a mode hook.
(subed-set-up-defaults): Moved from subed--init to a function so that
it can be in a mode hook.
(subed-mode): Use mode hooks to run code after derived modes have
initialized.
(item): Do not add to auto-mode-alist in the main file.
* tests/test-subed-ass.el (with-temp-ass-buffer): Use subed-ass-mode.
* tests/test-subed-common.el ("Trimming subtitles"): Use subed-srt-mode.
* tests/test-subed-srt.el (with-temp-srt-buffer): Use subed-srt-mode.
* tests/test-subed-vtt.el (with-temp-vtt-buffer): Use subed-vtt-mode.