Switch to subed-waveform-minor-mode, allow filter functions
* README.org (Features): Document `subed-waveform-minor-mode'. Move it
above more specialized features such as word timing data and
alignment.
* subed/subed-waveform.el: Update the comment to refer to
`subed-waveform-minor-mode'. Add paragraphs.
(subed-waveform): Define customize group.
(subed-waveform-ffmpeg-filter-args): Change the default to gray to
make it more likely to be somewhat visible. (I wonder if we can get
this to default to the user's foreground color...) Include the
documentation for setting it to a function.
(subed-show-waveform-p): Removed because I changed it to be a minor
mode like my other implementation. That way, we can just use the minor
mode variable `subed-waveform-minor-mode' to check if it's enabled or
not. Using a minor mode means that the code for enabling and disabling
is together in one place and follows standard conventions, and has a
bunch of extra niceties like announcing the status and letting people
add hooks. I know there are a bunch of other things in subed that
don't use minor modes (like the CPS overlay); I might switch them over
at some point.
(subed-enable-show-waveform): Removed, use `subed-waveform-minor-mode'
with a positive arg.
(subed-disable-show-waveform): Removed, use
`subed-waveform-minor-mode' with a negative arg.
(subed-waveform-remove): New convenience function since we want to
call this from the mode as well.
(subed-waveform-minor-mode): New function.
(subed-waveform-fancy-filter): Example fancy filter.
(subed-waveform--from-file): Accept fancy filter.
(subed-waveform--*): Generally renamed subed--waveform-* to
subed-waveform--* so that the nameless package can hide the prefix
more easily.
(subed-waveform-put-svg): Ignores arguments so that it can easily be
added to hooks.