1.2.12 bugfix for SRT numeric cue text
*
subed/subed-srt.el (subed--jump-to-subtitle-end):
Handle blank subtitles as a separate case so that
we can use subed-srt--regexp-separator to handle
subtitles with numbers in the cue text.

*
subed/subed-vtt.el (subed--jump-to-subtitle-end):
Update the comment.

* tests/test-subed-srt.el ("SRT"): Add a test case
for jumping to the end of a cue with a line that's
all numbers, and add a test for merging those cues
too.
* tests/test-subed-vtt.el ("VTT"): Add a test case
for the same scenario.

Thanks to rodrigomorales1 for the detailed bug
description!
https://github.com/sachac/subed/issues/71